requirements.txt
py-clob-client>=0.34
streamlit>=1.30.0
psycopg2-binary>=2.9
plotly>=5.20
aiohappyeyeballs==2.6.1
beautifulsoup4>=4.12
PyYAML==6.0.3
tabulate==0.10.0
pytest>=8.0
pytest-asyncio>=0.24
aioresponses>=0.7
ruff>=0.1
mypy>=1.0
types-PyYAML
types-tabulate
aiohttp==3.13.5
aiosignal==1.4.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.2.25
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
numpy==2.4.4
pandas==3.0.2
propcache==0.4.1
python-dateutil==2.9.0.post0
six==1.17.0
websockets==16.0
yarl==1.23.0
